✨Add lists of steps in one go
One of the things we've noticed develoing large test suites is that it
is very common to have sequences of three to four steps. For example,
here is a real-life example of a multi-step interaction that
represent navigation to a page, then a sub-page, and then setting some
filters:

```js
.step(App.visit('/users/preview/1'))
.step(Button('Fees/fines').click())
.step(Link('Create fee/fine').click())
.step(Select('Fee/fine owner*').select('testOwner'))
.step(Select('Fee/fine type*').select('testFineType'))
```

the fact that we have to add each step with a `step` invocation is not
only annoying to the author of the step, it more importantly obscures
what is happening to the reader of the test case. We can more simply
express a list of steps using a `stepList` method where we add all the
steps in a single shot:

```js
stepList([
  App.visit('/users/preview/1'),
  Button('Fees/fines').click(),
  Link('Create fee/fine').click(),
  Select('Fee/fine owner*').select('testOwner'),
  Select('Fee/fine type*').select('testFineType')
])
```

In order to make sure that the type-checking works out properly and
that the context that gets passed in a the top type-checks, we put a
constraint on these steps that there actions must be of type
`void`. That way we can be guaranteed that none of the original steps
contributes to the context.